requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
transformers>=4.30.0
accelerate>=0.20.0
google-cloud-vision>=3.4.0
pyyaml>=6.0
tqdm>=4.65.0
efficientnet-pytorch>=0.7.1
grad-cam>=1.4.6
plotly>=5.15.0
streamlit-option-menu>=0.3.6
streamlit-image-comparison>=0.0.4
albumentations>=1.3.0
